[rename]: 유틸 함수 파일명을 동사로 변경 (#199)
